Discount codes

+++ backend-parent/backend/src/main/java/cz/zorganizovano/backend/dao/OrderItemDao.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 public interface OrderItemDao extends JpaRepository<OrderItem, Long> {
 
     @Query("SELECT SUM(i.quantity * i.price) FROM OrderItem i WHERE i.order.id = :orderId GROUP BY i.order.id")
-    long getTotalOrderItemsPrice(long orderId);
+    double getTotalOrderItemsPrice(long orderId);
 
     List<OrderItem> findByOrder(Order order);
